requirements.txt
vkbottle==4.3.12
aiosqlite==0.18.0
python-dotenv==1.0.0